github-mcp-server/pyproject.toml
[project]
name = "github-mcp-server"
version = "0.1.0"
description = "GitHub API MCP Server built with FastMCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
